Fixing the chapter selection
The chapter selection in the APPs was wrong. E.g. selecting "Klasse A" only loads the chapters for A and not the chapters N, E and A. 
Also, the "Upgrade von E auf A" loads the chapters E and A. In this case, only A should be loaded. 
The same is true for the other buttons when selecting the "Klasse".